Rename the pgen module to stategraph.
Discussed with Lukas.
